all fixed

+++ FormMain/EventCreator.cs
@@ -1,6 +1,7 @@
 ﻿using System;
 using System.Collections.Generic;
 using System.Linq;
+using System.Numerics;
 using System.Text;
 using System.Threading.Tasks;
 
@@ -9,8 +10,8 @@ namespace FormMain
     public class EventCreatorEventArgs : EventArgs
     {
 
-        public long Number { get; set; }
-        public EventCreatorEventArgs(long number)
+        public BigInteger Number { get; set; }
+        public EventCreatorEventArgs(BigInteger number)
         {
             Number = number;
         }
